+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 Change Log
 ==========
 
-Version 0.?.? (??????????)
---------------------------
+Version 0.10.0 (2014.13.25)
+---------------------------
 
  * added 'toString' function
  * added 'jarjar' function